+++ d60.Cirqus.MongoDb/Events/MongoDbEventStore.cs
@@ -177,7 +177,7 @@ namespace d60.Cirqus.MongoDb.Events
             return _eventBatches.Find(criteria)
                 .SelectMany(b => b.Events)
                 .OrderBy(e => e.GlobalSequenceNumber)
-                .Where(e => e.GlobalSequenceNumber >= firstSeq && e.AggregateRootId == aggregateRootId)
+                .Where(e => e.SequenceNumber >= firstSeq && e.AggregateRootId == aggregateRootId)
                 .Select(MongoEventToEvent);
         }