+++ d60.Cirqus/Config/CoreConfigurationExtensions.cs
@@ -38,6 +38,17 @@ namespace d60.Cirqus.Config
         }
 
         /// <summary>
+        /// Enables an in-memory event cache that caches the most recently used events. <see cref="maxCacheEntries"/> specifies
+        /// the approximate number of events to be held in the cache
+        /// </summary>
+        public static void EnableCaching(this EventStoreConfigurationBuilder builder, int maxCacheEntries)
+        {
+            builder.Registrar
+                .Register<IEventStore>(context => new CachingEventStoreDecorator(context.Get<IEventStore>()),
+                    decorator: true);
+        }
+
+        /// <summary>
         /// Registers a <see cref="DefaultAggregateRootRepository"/> as the <see cref="IAggregateRootRepository"/> implementation. Since this is the
         /// default, there's no need to call this method explicitly.
         /// </summary>
@@ -74,14 +85,28 @@ namespace d60.Cirqus.Config
         /// Registers a <see cref="Views.ViewManagerEventDispatcher"/> to manage the given views. Can be called multiple times in order to register
         /// multiple "pools" of views (each will be managed by a dedicated worker thread).
         /// </summary>
+        public static void UseViewManagerEventDispatcher(this EventDispatcherConfigurationBuilder builder, int maxDomainEventsPerBatch, params IViewManager[] viewManagers)
+        {
+            InstallViewManagerEventDispatcherInternal(builder, viewManagers, maxDomainEventsPerBatch: maxDomainEventsPerBatch);
+        }
+
+        /// <summary>
+        /// Registers a <see cref="Views.ViewManagerEventDispatcher"/> to manage the given views. Can be called multiple times in order to register
+        /// multiple "pools" of views (each will be managed by a dedicated worker thread).
+        /// </summary>
         public static void UseViewManagerEventDispatcher(this EventDispatcherConfigurationBuilder builder, params IViewManager[] viewManagers)
         {
-            AddEventDispatcherRegistration(builder, context => new ViewManagerEventDispatcher(
-                context.Get<IAggregateRootRepository>(), 
-                context.Get<IEventStore>(), 
-                context.Get<IDomainEventSerializer>(),
-                context.Get<IDomainTypeNameMapper>(),
-                viewManagers));
+            InstallViewManagerEventDispatcherInternal(builder, viewManagers);
+        }
+
+        /// <summary>
+        /// Registers a <see cref="ViewManagerEventDispatcher"/> to manage the given views. Can be called multiple times in order to register
+        /// multiple "pools" of views (each will be managed by a dedicated worker thread). The event dispatcher will register itself with the
+        /// given <seealso cref="waitHandle"/>, allowing for optionally blocking until views have been updated to a certain point.
+        /// </summary>
+        public static void UseViewManagerEventDispatcher(this EventDispatcherConfigurationBuilder builder, ViewManagerWaitHandle waitHandle, int maxDomainEventsPerBatch, params IViewManager[] viewManagers)
+        {
+            InstallViewManagerEventDispatcherInternal(builder, viewManagers, waitHandle: waitHandle, maxDomainEventsPerBatch: maxDomainEventsPerBatch);
         }
 
         /// <summary>
@@ -90,6 +115,11 @@ namespace d60.Cirqus.Config
         /// given <seealso cref="waitHandle"/>, allowing for optionally blocking until views have been updated to a certain point.
         /// </summary>
         public static void UseViewManagerEventDispatcher(this EventDispatcherConfigurationBuilder builder, ViewManagerWaitHandle waitHandle, params IViewManager[] viewManagers)
+        {
+            InstallViewManagerEventDispatcherInternal(builder, viewManagers, waitHandle: waitHandle);
+        }
+
+        static void InstallViewManagerEventDispatcherInternal(EventDispatcherConfigurationBuilder builder, IViewManager[] viewManagers, ViewManagerWaitHandle waitHandle = null, int maxDomainEventsPerBatch = 0)
         {
             AddEventDispatcherRegistration(builder, context =>
             {
@@ -100,7 +130,15 @@ namespace d60.Cirqus.Config
                     context.Get<IDomainTypeNameMapper>(),
                     viewManagers);
 
-                waitHandle.Register(eventDispatcher);
+                if (maxDomainEventsPerBatch != 0)
+                {
+                    eventDispatcher.MaxDomainEventsPerBatch = maxDomainEventsPerBatch;
+                }
+
+                if (waitHandle != null)
+                {
+                    waitHandle.Register(eventDispatcher);
+                }
 
                 return eventDispatcher;
             });