+++ d60.Cirqus/Diagnostics/DiagnosticsConfigurationExtensions.cs
@@ -51,9 +51,9 @@ namespace d60.Cirqus.Diagnostics
                 _profiler = profiler;
             }
 
-            public void RecordAggregateRootGet(TimeSpan elapsed, string aggregateRootId)
+            public void RecordAggregateRootGet(TimeSpan elapsed, string aggregateRootId, Type aggregateRootType)
             {
-                _profiler.RecordAggregateRootGet(elapsed, aggregateRootId);
+                _profiler.RecordAggregateRootGet(elapsed, aggregateRootType, aggregateRootId);
             }
 
             public void RecordAggregateRootExists(TimeSpan elapsed, string aggregateRootId)
@@ -121,14 +121,20 @@ namespace d60.Cirqus.Diagnostics
             public AggregateRoot Get<TAggregateRoot>(string aggregateRootId, IUnitOfWork unitOfWork, long maxGlobalSequenceNumber = long.MaxValue, bool createIfNotExists = false)
             {
                 var stopwatch = Stopwatch.StartNew();
+                Type actualAggregateRootType = null;
+
                 try
                 {
-                    return _innnerAggregateRootRepository
+                    var aggregateRoot = _innnerAggregateRootRepository
                         .Get<TAggregateRoot>(aggregateRootId, unitOfWork, maxGlobalSequenceNumber, createIfNotExists);
+
+                    actualAggregateRootType = aggregateRoot.GetType();
+
+                    return aggregateRoot;
                 }
                 finally
                 {
-                    _operationProfiler.RecordAggregateRootGet(stopwatch.Elapsed, aggregateRootId);
+                    _operationProfiler.RecordAggregateRootGet(stopwatch.Elapsed, aggregateRootId, actualAggregateRootType ?? typeof(TAggregateRoot));
                 }
             }